AI_Service/requirements.txt
aio-pika==9.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiormq==6.9.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==4.0.3
attrs==25.4.0
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
exceptiongroup==1.3.1
fastapi==0.131.0
frozenlist==1.8.0
greenlet==3.3.2
grpcio==1.78.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.14
langchain-text-splitters==1.1.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
lxml==6.0.2
marshmallow==3.26.2
minio==7.2.20
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.2.6
orjson==3.11.7
ormsgpack==1.12.2
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pamqp==3.3.0
pika==1.3.2
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.5
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.8.0
PyMuPDF==1.27.1
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.0
redis==7.2.0
requests==2.32.5
requests-toolbelt==1.0.0
SQLAlchemy==2.0.46
starlette==0.52.1
tenacity==9.1.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
xxhash==3.6.0
yarl==1.22.0
zhipuai==2.1.5.20250825
zstandard==0.25.0
